Update logic for Checkout by User functionality - part 1

+++ backend/src/main/java/com/sagar/libraryapp/service/CheckoutService.java
@@ -2,8 +2,7 @@ package com.sagar.libraryapp.service;
 
 public interface CheckoutService {
 
-    void checkoutBook(String email, int bookId);
-
+    void checkoutBook(String email, long bookId);
     int checkedoutBooks(String email);
 
 }
